+++ CHANGELOG.md
@@ -161,6 +161,7 @@ No user-visible changes. This release was done solely to fix a problem with the 
   - `cider-read-and-eval` (`:`)
   - `cider-eval-defun-to-comment` (`;`)
   - `cider-pprint-eval-last-sexp-to-repl` (`f j e`)
+- [#3839](https://github.com/clojure-emacs/cider/pull/3839): Add jack-in support for ClojureCLR.
 
 ### Bugs fixed